Fix working of voice experience system.
